Skip all import tests when running without xattrs
- move specific tests in import_test.go to xattr_upgrade_test.go since
  they are specifically needed only when doing a rolling upgrade from
  inline -> xattr metadata
