+++ rest/importtest/main_test.go
@@ -19,6 +19,9 @@ import (
 )
 
 func TestMain(m *testing.M) {
+	if !base.TestUseXattrs() { // import tests only run if xattrs are enabled
+		return
+	}
 	ctx := context.Background() // start of test process
 	tbpOptions := base.TestBucketPoolOptions{MemWatermarkThresholdMB: 2048}
 	db.TestBucketPoolWithIndexes(ctx, m, tbpOptions)
